Fixed for no overflow

+++ include/mtkahypartypes.h
@@ -1,5 +1,6 @@
 #ifndef MTKAHYPAR_TYPEDEFS_H
 #define MTKAHYPAR_TYPEDEFS_H
+#include <stdint.h>
 
 #include <stddef.h>
 
@@ -81,8 +82,8 @@ typedef struct {
 
 typedef unsigned long int mt_kahypar_hypernode_id_t;
 typedef unsigned long int mt_kahypar_hyperedge_id_t;
-typedef int mt_kahypar_hypernode_weight_t;
-typedef int mt_kahypar_hyperedge_weight_t;
+typedef int64_t mt_kahypar_hypernode_weight_t;
+typedef int64_t mt_kahypar_hyperedge_weight_t;
 typedef int mt_kahypar_partition_id_t;
 
 /**
